Horrible Hack to share data classes with console app
The Console App (Dragon Commander) cannot reference the DragonAidLib
that we build because that library is only for Windows Store. We can't
convert it to a portable library (that is the right thing to do) because
the Azure Mobile Services SDK isn't a portable library yet.

Workaround: The console app now cross-compiles the source files from
DragonAidLib that it wants. They are linked files with occasional ifdefs
to remove that bindings that are needed for azure.