Added Languages to the data model
Added Languages as a first-level concept in the data model, following
the same pattern we use for weapons and skills.

Languages are technically skills, but I think it will be easier to call
out languages separately because characters have a different ranks for
speaking and reading/writing in a given language.